+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Change Log
 
+## 0.1.1
++ Every project has its spiders.
+
 ## 0.1.0
-+ modular spider
++ Modular spider.